CRUD Paciente y mejora en peticiones PUT

+++ ProyectoFinalBackEndI/src/main/java/com/ClinicaDH/ProyectoFinalBackEndI/persistance/models/Paciente.java
@@ -10,7 +10,7 @@ public class Paciente {
     @Id
     @SequenceGenerator(name = "paciente_sequence", sequenceName = "paciente_sequence")
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "paciente_sequence")
-    private Integer id;
+    private Long id;
 
     @Column
     private String nombre;
@@ -40,7 +40,7 @@ public class Paciente {
         this.domicilio = domicilio;
     }
 
-    public Paciente(Integer id, String nombre, String apellido, String dni, Date fechaIngreso, Domicilio domicilio) {
+    public Paciente(Long id, String nombre, String apellido, String dni, Date fechaIngreso, Domicilio domicilio) {
         this.id = id;
         this.nombre = nombre;
         this.apellido = apellido;
@@ -50,7 +50,7 @@ public class Paciente {
     }
 
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
